+++ tests/testthat/test-hardhat_parameters.R
@@ -5,6 +5,67 @@ test_that("errors when using an argument that do not exist", {
     "unused argument"
   )
 
+  expect_error(
+    fit <- tabnet_fit(x, y, config = tabnet_config(epochsas = 1)),
+    "unused argument"
+  )
+
+})
+
+test_that("merging parameters from config and dots works", {
+  
+  # we can merge non-atomic values on different variables (no comparison)
+  expect_no_error(
+    merged_conf <- tabnet:::merge_config_and_dots(
+      config = tabnet_config(optimizer = torch::optim_adamw),
+      loss = torch::nn_bce_loss)
+  )
+  expect_identical_modules(merged_conf$optimizer, torch::optim_adamw)
+  expect_identical_modules(merged_conf$loss, torch::nn_bce_loss)
+  
+  # we can merge non-atomic values on different variables while resolving optimizer
+  expect_no_error(
+    merged_conf <- tabnet:::merge_config_and_dots(
+      config = tabnet_config(optimizer = "adam"),
+      loss = torch::nn_bce_loss)
+  )
+  if(tabnet:::torch_has_optim_ignite()) {
+    expect_identical_modules(merged_conf$optimizer, torch::optim_ignite_adam)
+  } else {
+    expect_identical_modules(merged_conf$optimizer, torch::optim_adam)
+  }
+  expect_identical_modules(merged_conf$loss, torch::nn_bce_loss)
+  
+  # ... value wins over tabnet_config() value
+  expect_no_error(
+    merged_conf <- tabnet:::merge_config_and_dots(
+      config = tabnet_config(batch_size = 200),
+      batch_size = 400)
+  )
+  expect_identical(merged_conf$batch_size, 400)
+  
+  # ... value wins over tabnet_config() value for non-atomic parameter
+  expect_no_error(
+    merged_conf <- tabnet:::merge_config_and_dots(
+      config = tabnet_config(loss = torch::nn_cross_entropy_loss),
+      loss = torch::nn_bce_loss)
+  )
+  expect_identical_modules(merged_conf$loss, torch::nn_bce_loss)
+  
+  # NULL value get replaced and optimizer is resolved even lately
+  expect_no_error(
+    merged_conf <- tabnet:::merge_config_and_dots(
+      config = tabnet_config(loss = NULL, device = "cuda"),
+      loss = torch::nn_bce_loss, optimizer = "adam")
+  )
+  if(tabnet:::torch_has_optim_ignite()) {
+    expect_identical_modules(merged_conf$optimizer, torch::optim_ignite_adam)
+  } else {
+    expect_identical_modules(merged_conf$optimizer, torch::optim_adam)
+  }
+  expect_identical_modules(merged_conf$loss, torch::nn_bce_loss)
+
+  
 })
 
 test_that("pretrain and fit both work with early stopping", {
@@ -12,14 +73,14 @@ test_that("pretrain and fit both work with early stopping", {
   expect_message(
     pretrain <- tabnet_pretrain(attrix, attriy, epochs = 100, valid_split = 0.5, verbose=TRUE,
                                 early_stopping_tolerance=1e-7, early_stopping_patience=3, learn_rate = 0.2),
-    "Early stopping at epoch"
+    "Early-stopping at epoch"
   )
   expect_lt(length(pretrain$fit$metrics),100)
 
   expect_message(
     fit <- tabnet_fit(attrix, attriy, epochs = 100, valid_split = 0.5, verbose=TRUE,
                       early_stopping_tolerance=1e-7, early_stopping_patience=3, learn_rate = 0.2),
-    "Early stopping at epoch"
+    "Early-stopping at epoch"
   )
   expect_lt(length(fit$fit$metrics),100)
 
@@ -32,7 +93,7 @@ test_that("early stopping works wo validation split", {
     pretrain <- tabnet_pretrain(attrix, attriy, epochs = 100, verbose=TRUE,
                                 early_stopping_monitor="train_loss",
                                 early_stopping_tolerance=1e-7, early_stopping_patience=3, learn_rate = 0.2),
-    "Early stopping at epoch"
+    "Early-stopping at epoch"
   )
   expect_lt(length(pretrain$fit$metrics),100)
 
@@ -48,7 +109,7 @@ test_that("early stopping works wo validation split", {
     fit <- tabnet_fit(attrix, attriy, epochs = 200, verbose=TRUE,
                       early_stopping_monitor="train_loss",
                       early_stopping_tolerance=1e-7, early_stopping_patience=3, learn_rate = 0.3),
-    "Early stopping at epoch"
+    "Early-stopping at epoch"
   )
   expect_lt(length(fit$fit$metrics),200)
 
@@ -98,6 +159,23 @@ test_that("step scheduler works", {
 
 })
 
+test_that("configuring optimizer works", {
+
+  expect_no_error(
+    fit <- tabnet_fit(x, y, epochs = 3, config = tabnet_config(optimizer = "adam"))
+  )
+
+  expect_no_error(
+    fit <- tabnet_fit(x, y, epochs = 3, optimizer = torch::optim_adamw)
+  )
+
+  expect_error(
+    fit <- tabnet_fit(x, y, epochs = 3, optimizer = "adamw"),
+    "Currently only \"adam\" is supported"
+  )
+
+})
+
 test_that("reduce_on_plateau scheduler works", {
 
   expect_no_error(
@@ -108,7 +186,7 @@ test_that("reduce_on_plateau scheduler works", {
   expect_error(
     fit <- tabnet_fit(x, y, epochs = 3, lr_scheduler = "multiplicative",
                       lr_decay = 0.1, step_size = 1),
-    "only the 'step' and 'reduce_on_plateau' scheduler"
+    "only the \"step\" and \"reduce_on_plateau\" scheduler"
   )
 
   sc_fn <- function(optimizer) {
@@ -139,21 +217,6 @@ test_that("fit uses config parameters mix from config= and ...", {
 
 })
 
-test_that("fit works with entmax mask-type", {
-
-  rec <- recipe(EnvironmentSatisfaction ~ ., data = attrition[ids, ]) %>%
-    step_normalize(all_numeric(), -all_outcomes())
-
-  expect_no_error(
-    tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
-                      config = tabnet_config( mask_type="entmax"))
-  )
-  expect_no_error(
-    predict(tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
-                       config = tabnet_config( mask_type="entmax")), attrition)
-  )
-
-})
 
 test_that("fit raise an error with non-supported mask-type", {
 
@@ -162,7 +225,7 @@ test_that("fit raise an error with non-supported mask-type", {
   expect_error(
     tabnet_fit(rec, attrition, epochs = 1, valid_split = 0.25, verbose = TRUE,
                       config = tabnet_config( mask_type="max_entropy")),
-    regexp = "either 'sparsemax' or 'entmax' as"
+    regexp = "either \"sparsemax\", \"sparsemax15\", \"entmax\" or \"entmax15\" as"
   )
 
 })